refactor(core): move `checkbox` to `input` category
Move checkbox to input to match its use case

BREAKING CHANGE: checkbox component are now inside the `core/Input` dir
