+++ packages/lib/src/App/app-comp.vue
@@ -303,7 +303,7 @@
   import IconButton from '@/components/Button/icon-button.vue'
   import CardBody from '@/components/Card/card-body.vue'
   import Card from '@/components/Card/card.vue'
-  import Checkbox from '@/components/Checkbox/checkbox.vue'
+  import Checkbox from '@/components/Input/checkbox.vue'
   import ChipGroup from '@/components/Chip/chip-group.vue'
   import Chip from '@/components/Chip/chip.vue'
   import Divider from '@/components/Divider/divider.vue'
@@ -311,8 +311,8 @@
   import Frame from '@/components/Frame/frame.vue'
   import BlockImage from '@/components/Image/block-image.vue'
   import SquareImage from '@/components/Image/square-image.vue'
-  import NumberInput from '@/components/Input/number-input.vue'
-  import TextInput from '@/components/Input/text-input.vue'
+  import NumberInput from '@/components/Input/number.vue'
+  import TextInput from '@/components/Input/text.vue'
   import List from '@/components/List/list.vue'
   import ColorBlock from '@/components/Misc/color-block.vue'
   import HybridIcon from '@/components/Misc/hybrid-icon.vue'
