+++ packages/lib/src/App/layout-inherit.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
   import Box from '@/components/Box/box.vue'
   import Button from '@/components/Button/button.vue'
-  import Checkbox from '@/components/Checkbox/checkbox.vue'
+  import Checkbox from '@/components/Input/checkbox.vue'
   import Form from '@/components/Form/form.vue'
   import Layout from '@/components/Layout/layout.vue'
   import List from '@/components/List/list.vue'
@@ -12,7 +12,7 @@
   import { customRef } from '@/ref/custom-ref'
   import { useFocusLock } from '@/ref/use-focus-lock'
   import { ref } from 'vue'
-  import TextInput from '@/components/Input/text-input.vue'
+  import TextInput from '@/components/Input/text.vue'
   import { useTooltip } from '@/ref/use-tooltip'
 
   const isDark = ref(true)
