--- packages/lib/src/components/Checkbox/index.ts
@@ -1,1 +1,0 @@
-export { default as Checkbox } from './checkbox.vue'
