+++ packages/lib/src/components/Input/char-counter.vue
