+++ packages/lib/src/components/Input/index.ts
@@ -1,12 +1,19 @@
-import Number from './number-input.vue'
-import Text from './text-input.vue'
+import Checkbox from './checkbox.vue'
+import Number from './number.vue'
+import Text from './text.vue'
 
 type InputComponent = typeof Text & {
   Number: typeof Number
   Text: typeof Text
+  Checkbox: typeof Checkbox
 }
 
-export default Object.assign({}, Text, { Text, Number }) as InputComponent
+export default Object.assign({}, Text, {
+  Text,
+  Number,
+  Checkbox
+}) as InputComponent
 
-export { default as NumberInput } from './number-input.vue'
-export { default as TextInput } from './text-input.vue'
+export { default as Checkbox } from './checkbox.vue'
+export { default as NumberInput } from './number.vue'
+export { default as TextInput } from './text.vue'
